fix(moon): revamp building all docker projects
Up until now, everything was built within docker builds. It turns out, much of
the work was duplicated, as exactly the same builds would happen in parallel :((
It seems moon is not really a good tool for docker-heavy workflows (or I'm just
holding it wrong). Anyways, now I'm building most of the things on device, and
transferring the stuff that was built to containers (except rust server backend
for the blog). This improves building times.